better separation of util functions
