+++ src/type-renderer/cf-default-content-type-renderer.ts
