+++ src/type-renderer/cf-render-prop-link.ts
@@ -1,7 +1,7 @@
 import {Field} from 'contentful';
 import {renderTypeGeneric, renderTypeUnion} from '../renderer';
 import {RenderContext} from './render-types';
-import {linkContentTypeValidations} from '../utils';
+import {linkContentTypeValidations} from '../cf-extract-validation';
 
 export const renderPropLink = (field: Pick<Field, 'validations' | 'linkType'>, context: RenderContext) => {
     const linkContentType = (field: Pick<Field, 'validations'>, context: RenderContext): string => {
