+++ src/type-renderer/index.ts
@@ -5,4 +5,4 @@ export {renderPropLink} from './cf-render-prop-link';
 export {renderRichText} from './cf-render-prop-richtext';
 export {createDefaultContext} from './default-context';
 export {defaultRenderers} from './cf-default-renderers';
-export {DefaultContentTypeRenderer} from './content-type-renderer';
+export {DefaultContentTypeRenderer} from './cf-default-content-type-renderer';
